+++ lib/os/mutex.c
@@ -14,7 +14,7 @@ static struct k_mutex *get_k_mutex(struct sys_mutex *mutex)
 	struct z_object *obj;
 
 	obj = z_object_find(mutex);
-	if (obj == NULL || obj->type != K_OBJ_SYS_MUTEX) {
+	if (obj == NULL || obj->type != (uint8_t)K_OBJ_SYS_MUTEX) {
 		return NULL;
 	}
